+++ js/main.js
@@ -1,5 +1,6 @@
-import {getPicture} from './util.js';
+import {getPictures} from './data.js';
 
 // 1. Функция получения рандомных картинок и комментариев;
-
-getPicture();
+const randomPicture = getPictures();
+// eslint-disable-next-line no-console
+console.log (randomPicture);
