+++ js/util.js
@@ -1,5 +1,3 @@
-import {PICTURE_COUNT, AVATAR_COUNT, LIKES_COUNT, COMMENT_COUNT, COMMENT_LINES, NAMES, DESCRIPTIONS} from './data.js';
-
 export const getRandomInteger = (a, b) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
@@ -17,29 +15,3 @@ export const createIdGenerator = () => {
     return lastGeneratedId;
   };
 };
-
-
-// 1. Функции получения рандомных картинок и комментариев.
-const generateCommentId = createIdGenerator();
-
-const getRandomCommentLines = () => Array.from({length: getRandomInteger(1, 2)}, () => getRandomArrayElement(COMMENT_LINES)).join(' ');
-
-const createComments = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, AVATAR_COUNT)}.svg`,
-  message: getRandomCommentLines(),
-  name: getRandomArrayElement(NAMES)
-});
-
-const createPictures = (index) => ({
-  id: index,
-  url: `photos/${index}.jpg`,
-  description: getRandomArrayElement(DESCRIPTIONS),
-  likes: getRandomInteger(15, LIKES_COUNT),
-  comments: Array.from({length: getRandomInteger(0, COMMENT_COUNT)}, createComments)
-});
-
-const getPicture = () => Array.from({length: PICTURE_COUNT}, (_, pictureIndex) => createPictures(pictureIndex + 1));
-
-export {getPicture};
-
